requirements.txt
# Requires Python >= 3.9 (uses builtin generic types: list[], dict[])
PyOpenGL>=3.1.7
glfw>=2.7.0
numpy>=1.24.0
Pillow>=10.1.0
